Fix: Fast Calculation with 'parallel'

+++ sbylib/source/sbylib/utils/FpsCounter.d
@@ -15,7 +15,7 @@ class FpsCounter(uint N) {
         timeCounter.start();
     }
 
-    long getFPS() {
+    float getFPS() {
         if (timeCounter.averageTime == 0) return 0;
         return 1000 / timeCounter.averageTime;
     }
